requirements.txt
fpdf2==2.7.9
art==6.2
